+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 0.31.1
+## Features Added
+  - Many improvements to `roboto.ai.Chat` and `roboto chat start`
+
+## Bugs Fixed
+  - `roboto actions init` no longer prompts to delete and re-download a previously cached Action template (if exists). This should always happen to ensure new Actions are making use of the latest template.
+
 # 0.31.0
 ## Features Added
   - Add `RobotoSearch::from_env` classmethod for convenience and consistency with other Roboto resource factories.